requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Training and visualization
tensorboard>=2.12.0
matplotlib>=3.7.0
pyyaml>=6.0

# Data processing
scikit-learn>=1.2.0

# Optional: For better performance
# tqdm>=4.65.0

# Development (optional)
# jupyter>=1.0.0
# ipython>=8.12.0
